Разделение скриптов на модули
