+++ js/params.js
@@ -1,0 +1,16 @@
+// Количество постов/фото
+const POSTS_QUANTITY = 25;
+
+// Интервал кол-ва лайков
+const LIKES_RANGE = {
+  min: 15,
+  max: 200
+};
+
+// Интервал кол-ва комментариев
+const COMMENTS_RANGE = {
+  min: 0,
+  max: 200
+};
+
+export {POSTS_QUANTITY, LIKES_RANGE, COMMENTS_RANGE};
